test: move to test/regression/issue/29346.test.ts, self-contained
Addresses review: regression tests for numbered issues belong in
test/regression/issue/, and the previous refactor of addr32.test.ts
pulled its shared-library compilation into a top-level beforeAll that
would fail on non-Linux platforms before test.skipIf had a chance to
run. Revert addr32.test.ts / addr32.c to their pre-PR state and make
the issue #29346 regression test fully self-contained via tempDir.
